requirements.txt
boto3==1.34.117
contractions==0.1.73
fastapi==0.111.0
joblib==1.4.2
mangum==0.17.0
neptune==1.10.4
nltk==3.8.1
numpy
pandas==2.2.2
python-dotenv==1.0.1
PyYAML==6.0.1
scikit_learn==1.4.2
Unidecode==1.3.8
uvicorn==0.30.1
